combine decorator tests
